Vital merge case that I was missing!
I thought that non-sibling, concurrent atoms could never have a conflict, and that might have been right in the original version… but not as much when you have priority atoms. I have now designed a canonical merge function that returns the correct order for ANY two atoms, though it’s expensive in this very edge case. (Still O(weave), but with a high constant factor.)